Change the folders structure to resemble the one of HEENS.
Create two different projects for ZedBoard and Zynq ZC706:
they include different files for the HDMI connection
and the events package that defines the constants for the number of neurons and the name of the board.

HDMI connection now works with both platforms.

Change has been made for the conversion RGB to YCbCr,
the I2C communication,
the registers writing for ADV7511 that manages the HDMI
and HDMI outputs: ZedBoard don't use DDR (Double Data Rate) and only 8 HDMI data pins anymore, but the whole 16 pins it has.

Zynq ZC706 possesses 24 pins and we can communicate directly in RGB color format,
the HDMI chip (ADV7511) has been configured accordingly, as well as the HDMI outputs manager.
This platform also uses a I2C Mux, that we need to configure first via I2C to enable the HDMI chip as slave.